Move MSM bench to a different file
